refactor: 패키지 위치 수정
- rest, view 패키지로 이동
